Use long id.

+++ Chemiklani.DAL/Entities/AppUserClaim.cs
@@ -3,7 +3,7 @@ using Microsoft.AspNet.Identity.EntityFramework;
 
 namespace Chemiklani.DAL.Entities
 {
-    public class AppUserClaim:IdentityUserClaim<int>
+    public class AppUserClaim:IdentityUserClaim<long>
     {
     }
 }
